Made some more changes - should work now
