+++ src/components/Main.tsx
@@ -38,6 +38,106 @@ interface StateProps {
 type Props = OwnProps & StateProps
 
 const Main: React.FC<Props> = (props: Props) => {
+  interface SortedActivity extends Activity {
+    startDateTime: Date
+    endDateTime: Date
+  }
+
+  const createActivityArray = (activities: Activity[]): SortedActivity[] => {
+    const result: SortedActivity[] = []
+
+    // Loop over all activities to create items for in the home screen
+    activities.forEach((activity) => {
+      if (activity.day) {
+        // If the activity is a single event
+
+        // Split date to values
+        const dateArray = activity.day.date.split('-')
+        // Split start time to values
+        const startTimeArray = activity.day.startTime.split(':')
+        // Create start date time object with values from activity
+        const startDateTime = new Date(
+          Number(dateArray[0]),
+          Number(dateArray[1]) - 1,
+          Number(dateArray[2]),
+          Number(startTimeArray[0]),
+          Number(startTimeArray[1]),
+        )
+        // Split end time to values
+        const endTimeArray = activity.day.endTime.split(':')
+        // Create end date time object with values from activity
+        const endDateTime = new Date(
+          Number(dateArray[0]),
+          Number(dateArray[1]) - 1,
+          Number(dateArray[2]),
+          Number(endTimeArray[0]),
+          Number(endTimeArray[1]),
+        )
+        // Push result to array as sorted activity
+        result.push({
+          ...activity,
+          startDateTime,
+          endDateTime,
+        })
+      } else if (activity.days) {
+        // If the activity is repeated weekly
+
+        // Filter all activity entries where the day is not defined and loop over all non-empty values
+        activity.days
+          .filter((day) => day.startTime !== '' && day.endTime !== '')
+          .forEach((day) => {
+            // Get the weekday number by string value
+            const weekday = getDayByString(day.name)
+
+            // Set amount of days to display
+            const amountOfDays = 14
+            // For each day
+            for (let i = 0; i < amountOfDays; i++) {
+              const date = new Date()
+              date.setSeconds(0)
+              date.setMilliseconds(0)
+              // Split start time to values
+              const startTimeArray = day.startTime.split(':')
+              // Split end time to values
+              const endTimeArray = day.endTime.split(':')
+
+              // Create start date time object with values from activity
+              const startDateTime = new Date(date)
+              startDateTime.setDate(startDateTime.getDate() + i)
+              startDateTime.setHours(Number(startTimeArray[0]))
+              startDateTime.setMinutes(Number(startTimeArray[1]))
+              // Create end date time object with values from activity
+              const endDateTime = new Date(date)
+              endDateTime.setDate(endDateTime.getDate() + i)
+              endDateTime.setHours(Number(endTimeArray[0]))
+              endDateTime.setMinutes(Number(endTimeArray[1]))
+
+              // Check if weekday equals to activity day, if so continue, else stop
+              if (weekday === startDateTime.getDay())
+                if (endDateTime >= new Date()) {
+                  // Check if end date is not passed, if so push, else change to next week
+                  result.push({
+                    ...activity,
+                    startDateTime,
+                    endDateTime,
+                  })
+                } else if (endDateTime <= date) {
+                  startDateTime.setDate(startDateTime.getDate() + amountOfDays)
+                  endDateTime.setDate(endDateTime.getDate() + amountOfDays)
+                  result.push({
+                    ...activity,
+                    startDateTime,
+                    endDateTime,
+                  })
+                }
+            }
+          })
+      }
+    })
+    return result
+  }
+
+  // check if activities are loaded, display a loader if it's still loading
   if (!isLoaded(props.activities))
     return (
       <Loader
@@ -47,7 +147,9 @@ const Main: React.FC<Props> = (props: Props) => {
         text="Activiteiten laden..."
       />
     )
+  // check if there are any, display 'Geen activiteiten gevonden.' if there ain't no activities
   else if (isEmpty(props.activities)) return <p>Geen activiteiten gevonden.</p>
+  // check if categories are loaded, display a loader if it's still loading
   else if (!isLoaded(props.categories))
     return (
       <Loader
@@ -57,6 +159,7 @@ const Main: React.FC<Props> = (props: Props) => {
         text="Categorieën laden..."
       />
     )
+  // check if organisers are loaded, display a loader if it's still loading
   else if (!isLoaded(props.organisers))
     return (
       <Loader
@@ -66,6 +169,7 @@ const Main: React.FC<Props> = (props: Props) => {
         text="Organisatoren laden..."
       />
     )
+  // check if registrations are loaded, display a loader if it's still loading
   else if (!isLoaded(props.registrations))
     return (
       <Loader
@@ -78,90 +182,23 @@ const Main: React.FC<Props> = (props: Props) => {
   else {
     const { activities, registrations } = props
 
-    interface SortedActivity extends Activity {
-      startDateTime: Date
-      endDateTime: Date
-    }
-    let sortedActivities: SortedActivity[] = []
+    let sortedActivities = createActivityArray(activities)
 
-    activities.forEach((activity) => {
-      if (activity.day) {
-        const dateArray = activity.day.date.split('-')
-        const startTimeArray = activity.day.startTime.split(':')
-        const startDateTime = new Date(
-          Number(dateArray[0]),
-          Number(dateArray[1]) - 1,
-          Number(dateArray[2]),
-          Number(startTimeArray[0]),
-          Number(startTimeArray[1]),
-        )
-        const endTimeArray = activity.day.endTime.split(':')
-        const endDateTime = new Date(
-          Number(dateArray[0]),
-          Number(dateArray[1]) - 1,
-          Number(dateArray[2]),
-          Number(endTimeArray[0]),
-          Number(endTimeArray[1]),
-        )
-        sortedActivities.push({
-          ...activity,
-          startDateTime,
-          endDateTime,
-        })
-      } else if (activity.days)
-        activity.days
-          .filter((day) => day.startTime !== '' && day.endTime !== '')
-          .forEach((day) => {
-            const weekday = getDayByString(day.name)
-
-            const amountOfDays = 14
-            for (let i = 0; i < amountOfDays; i++) {
-              const date = new Date()
-              date.setSeconds(0)
-              date.setMilliseconds(0)
-              const startTimeArray = day.startTime.split(':')
-              const endTimeArray = day.endTime.split(':')
-
-              const startDateTime = new Date(date)
-              startDateTime.setDate(startDateTime.getDate() + i)
-              startDateTime.setHours(Number(startTimeArray[0]))
-              startDateTime.setMinutes(Number(startTimeArray[1]))
-              const endDateTime = new Date(date)
-              endDateTime.setDate(endDateTime.getDate() + i)
-              endDateTime.setHours(Number(endTimeArray[0]))
-              endDateTime.setMinutes(Number(endTimeArray[1]))
-
-              if (weekday === startDateTime.getDay()) {
-                if (endDateTime >= new Date())
-                  sortedActivities.push({
-                    ...activity,
-                    startDateTime,
-                    endDateTime,
-                  })
-                else if (endDateTime <= date) {
-                  startDateTime.setDate(startDateTime.getDate() + amountOfDays)
-                  endDateTime.setDate(endDateTime.getDate() + amountOfDays)
-                  sortedActivities.push({
-                    ...activity,
-                    startDateTime,
-                    endDateTime,
-                  })
-                }
-              }
-            }
-          })
-    })
-
+    // sort the activities by time
     sortedActivities.sort((a, b) => {
       if (a.startDateTime < b.startDateTime) return -1
       if (a.startDateTime > b.startDateTime) return 1
       else return 0
     })
 
+    // display only activities that haven't passed yet
     sortedActivities = sortedActivities.filter(
       (_activity) => _activity.endDateTime >= new Date(),
     )
 
+    // filter by searchbar
+    // toLowerCase to remove case sensitive searches
+    // searchbar works with name, room, descrition and createdBy
     if (props.filters.search)
       sortedActivities = sortedActivities.filter(
         (_activity) =>
@@ -188,8 +225,10 @@ const Main: React.FC<Props> = (props: Props) => {
       props.filters.spiritueel ||
       props.filters.sociaal
     ) {
-      // we waren hier
+      // TODO: continue here for multiple filters functionality
     }
+
+    // display only activities with the 'beweging' category
     if (props.filters.beweging)
       sortedActivities = sortedActivities.filter((_activity) =>
         props.categories.find(
@@ -199,6 +238,7 @@ const Main: React.FC<Props> = (props: Props) => {
         ),
       )
 
+    // display only activities with the 'creatief' category
     if (props.filters.creatief)
       sortedActivities = sortedActivities.filter((_activity) =>
         props.categories.find(
@@ -207,6 +247,8 @@ const Main: React.FC<Props> = (props: Props) => {
             _activity.category.split('/')[1] === _category.id,
         ),
       )
+
+    // display only activities with the 'kinderen' category
     if (props.filters.kinderen)
       sortedActivities = sortedActivities.filter((_activity) =>
         props.categories.find(
@@ -215,6 +257,8 @@ const Main: React.FC<Props> = (props: Props) => {
             _activity.category.split('/')[1] === _category.id,
         ),
       )
+
+    // display only activities with the 'sociaal' category
     if (props.filters.sociaal)
       sortedActivities = sortedActivities.filter((_activity) =>
         props.categories.find(
@@ -223,6 +267,8 @@ const Main: React.FC<Props> = (props: Props) => {
             _activity.category.split('/')[1] === _category.id,
         ),
       )
+
+    // display only activities with the 'spiritueel' category
     if (props.filters.spiritueel)
       sortedActivities = sortedActivities.filter((_activity) =>
         props.categories.find(
@@ -231,6 +277,8 @@ const Main: React.FC<Props> = (props: Props) => {
             _activity.category.split('/')[1] === _category.id,
         ),
       )
+
+    // display only activities with the 'taal' category
     if (props.filters.taal)
       sortedActivities = sortedActivities.filter((_activity) =>
         props.categories.find(
@@ -240,6 +288,7 @@ const Main: React.FC<Props> = (props: Props) => {
         ),
       )
 
+    // display only activities that are registered to
     if (props.filters.mijn)
       sortedActivities = sortedActivities.filter((_activity) =>
         registrations.find(
@@ -249,6 +298,7 @@ const Main: React.FC<Props> = (props: Props) => {
         ),
       )
 
+    // display only activities that happen once
     if (props.filters.speciaal)
       sortedActivities = sortedActivities.filter((_activity) => {
         if (_activity.day) return true
@@ -257,6 +307,7 @@ const Main: React.FC<Props> = (props: Props) => {
 
     return (
       <StyledTimeline>
+        {/* map activities and find the right category for the activty */}
         {sortedActivities.map(
           (activity: SortedActivity, i: number): React.ReactNode => {
             const category = props.categories.find(
@@ -266,6 +317,8 @@ const Main: React.FC<Props> = (props: Props) => {
             let displayMonth = true
 
             // See if activity is first of the day
+
+            // check if previous item in the list shows a date, don't display if it does
             if (
               sortedActivities[i - 1]?.startDateTime.getFullYear() ===
                 activity.startDateTime.getFullYear() &&
@@ -276,6 +329,7 @@ const Main: React.FC<Props> = (props: Props) => {
             )
               displayDay = false
 
+            // check if previous item in the list shows a month, don't display if it does
             if (
               sortedActivities[i - 1]?.startDateTime.getFullYear() ===
                 activity.startDateTime.getFullYear() &&
