+++ src/components/header/dropdown/types.ts
@@ -1,0 +1,9 @@
+export interface Props {
+  closeDropdown: Function
+  children: React.ReactNode
+  open: boolean
+}
+
+export interface StyledDropdownProps {
+  open: boolean
+}
