--- src/components/header/index.ts
@@ -1,2 +1,0 @@
-export { default as Dropdown } from './Dropdown'
-export { default as Header } from './Header'
