+++ src/components/index.ts
@@ -1,15 +1,14 @@
 export { default as Activity } from './activity'
-export * from './activity'
 export { default as Modal } from './modal'
-export * from './modal'
+export { default as Loader } from './loader'
+export { default as Header } from './header'
+export { default as Backbutton } from './back-button'
 
-export { default as Backbutton } from './Backbutton'
 export { default as SearchBar } from './SearchBar'
 export { default as FilterItem } from './FilterItem'
 export { default as Main } from './Main'
 export { default as Icon } from './Icon'
 export { default as BackgroundLogo } from './BackgroundLogo'
-export { default as Loader } from './Loader'
-export * from './toast'
-export * from './header'
 export { default as Filter } from './filter'
+
+export * from './toast'
