+++ src/components/loader/types.ts
@@ -1,0 +1,17 @@
+export interface Props {
+  height?: string
+  width?: string
+  scale?: number
+  color?: string
+  text?: string
+  type?: string
+  fontSize?: string
+}
+
+export interface StyledLoaderProps {
+  readonly height: string
+  readonly width: string
+  readonly scale: number
+  readonly color: string
+  readonly fontSize: string
+}
