+++ src/components/toast/close-button/index.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react'
+import { IconContainer, Close as CloseIcon } from 'src/icons'
+import { StyledContainer } from './styles'
+import { Props } from './types'
+
+const CloseButton: React.FC<Props> = (props: Props) => {
+  return (
+    <StyledContainer onClick={props.closeToast}>
+      <IconContainer>
+        <CloseIcon />
+      </IconContainer>
+    </StyledContainer>
+  )
+}
+
+export default CloseButton
