+++ src/components/toast/close-button/types.ts
@@ -1,0 +1,3 @@
+export interface Props {
+  closeToast?: any // eslint-disable-line @typescript-eslint/no-explicit-any
+}
