+++ src/components/toast/index.ts
@@ -1,2 +1,2 @@
-export { default as CloseButton } from './CloseButton'
+export { default as CloseButton } from './close-button'
 export * from './style'
