+++ src/pages/settings/index.tsx
@@ -1,209 +1,37 @@
 import * as React from 'react'
-import styled from 'styled-components'
+import { Link } from 'react-router-dom'
+import { connect } from 'react-redux'
 import { Header, BackgroundLogo, Icon } from 'src/components'
-import { Link, RouteComponentProps } from 'react-router-dom'
 import NameIcon from 'src/assets/general/icon_settings_name.svg'
 import MailIcon from 'src/assets/general/icon_settings_mail.svg'
 import PhoneIcon from 'src/assets/general/icon_settings_phone.svg'
-import HelpIcon from 'src/assets/general/icon_settings_help.svg'
+// import HelpIcon from 'src/assets/general/icon_settings_help.svg'
 import GlobeIcon from 'src/assets/general/icon_settings_globe.svg'
 import PrivacyIcon from 'src/assets/general/icon_settings_privacy.svg'
-import { layout, colors } from 'src/styles'
-import { connect } from 'react-redux'
 import { RootState } from 'src/redux/store'
 import { fireStore } from 'src/utils/firebase'
-import { Profile } from 'src/types/react-redux-firebase'
-
-interface StyledProps {
-  focus: boolean
-}
-
-interface OwnProps {}
-
-interface StateProps {
-  isLoggedIn: boolean
-  profile: Profile
-}
-
-type Props = OwnProps & StateProps & RouteComponentProps
-
-const StyledWrapper = styled.div`
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-  margin-top: 10px;
-  padding: 0px 15px;
-  max-width: 400px;
-  margin: auto;
-
-  a {
-    text-decoration: none;
-  }
-`
-
-const StyledTitle = styled.h2`
-  color: ${colors.colors.darkgrey};
-  font-size: 16px;
-  margin-bottom: 20px;
-  align-self: flex-start;
-`
-
-const StyledContainer = styled.div<StyledProps>`
-  position: relative;
-  height: 48px;
-  width: 100%;
-  border-radius: 99999px;
-  box-shadow: ${colors.shadows.input};
-  background: ${colors.colors.white};
-  margin-bottom: 16px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  overflow: hidden;
-
-  > div {
-    z-index: 1;
-    transition: all 0.2s ease-in-out;
-    opacity: ${(props): number => (props.focus ? 0.75 : 0.25)};
-  }
-
-  > div:last-of-type {
-    padding: 24px;
-    opacity: 1;
-  }
-`
-
-const StyledInput = styled.input`
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  flex-grow: 1;
-  background: none;
-  border: none;
-  border-radius: 99999px;
-  outline: none;
-  padding: 0 ${layout.unit * 3}px;
-  font-size: 16px;
-  color: ${colors.colors.black};
-  transition: all 0.2s ease-in-out;
-
-  &:focus {
-    box-shadow: inset ${colors.shadows.default};
-  }
-
-  &::placeholder {
-    color: ${colors.colors.black};
-    opacity: 0.5;
-  }
-`
-
-const StyledButton = styled.button`
-  background: ${colors.colors.orange};
-  color: ${colors.colors.white};
-  text-decoration: none;
-  border: none;
-  border-radius: 9999px;
-  width: 190px;
-  height: 30px;
-  font-size: 16px;
-  font-weight: bold;
-  margin: 15px 0px;
-`
-
-const StyledHorizontalRule = styled.hr`
-  background-color: #d3d3d3;
-  border: none;
-  height: 1px;
-  margin: 15px 0 25px;
-  width: 100%;
-`
-
-const StyledToggleContainer = styled.div`
-  margin-bottom: 10px;
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-`
-
-const StyledToggle = styled.label`
-  position: relative;
-  display: inline-block;
-  width: 50px;
-  height: 28px;
-
-  .slider {
-    position: absolute;
-    cursor: pointer;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    background-color: ${colors.colors.lightgrey};
-    transition: 0.2s;
-    border-radius: 34px;
-    width: 50px;
-
-    &::before {
-      position: absolute;
-      content: '';
-      height: 21px;
-      width: 21px;
-      left: 4px;
-      bottom: 4px;
-      background-color: white;
-      transition: 0.2s;
-      border-radius: 50%;
-    }
-  }
-`
-
-const StyledCheckbox = styled.input`
-  opacity: 0;
-  width: 0;
-  height: 0;
-  border-radius: 9999px;
-
-  &:checked + .slider {
-    background-color: ${colors.colors.orange};
-  }
-
-  &:focus + .slider {
-    box-shadow: 0 0 1px #2196f3;
-  }
-
-  &:checked + .slider:before {
-    transform: translateX(21px);
-  }
-`
-const StyledToggleLabel = styled.span`
-  color: ${colors.colors.darkgrey};
-  margin: 3px 0 0 0px;
-
-  a {
-    text-decoration: none;
-  }
-`
-
-const StyledBottomLabel = styled.div`
-  color: ${colors.colors.orange};
-  display: flex;
-  margin-bottom: 10px;
-
-  p {
-    margin-left: 5px;
-  }
-`
+import { Props, OwnProps, StateProps } from './types'
+import {
+  StyledWrapper,
+  StyledTitle,
+  StyledContainer,
+  StyledInput,
+  StyledButton,
+  StyledHorizontalRule,
+  StyledToggleContainer,
+  StyledToggleLabel,
+  StyledToggle,
+  StyledCheckbox,
+  StyledBottomLabel,
+} from './styles'
 
 const Settings: React.FC<Props> = (props: Props) => {
   const [focused1, setFocused1] = React.useState(false)
   const [focused2, setFocused2] = React.useState(false)
   const [focused3, setFocused3] = React.useState(false)
-  const [name, setName] = React.useState(props.profile.displayName)
-  const [email, setEmail] = React.useState(props.profile.email)
-  const [phone, setPhone] = React.useState(props.profile.phoneNumber)
+  const [name, setName] = React.useState(props.profile.displayName || '')
+  const [email, setEmail] = React.useState(props.profile.email || '')
+  const [phone, setPhone] = React.useState(props.profile.phoneNumber || '')
 
   const handleChange = (event: React.ChangeEvent<HTMLInputElement>): void => {
     if (event.target.name === 'name') setName(event.target.value)
@@ -212,19 +40,20 @@ const Settings: React.FC<Props> = (props: Props) => {
   }
 
   const saveSettings = (): void => {
-    // Save naar firebase
-    fireStore.collection('users').doc(props.profile.uid).update({
+    // Save settings to firestore
+    fireStore.collection('users').doc(props.profile.id).update({
       displayName: name,
       email,
       phone,
     })
   }
 
-  let inputElement1: any
-  let inputElement2: any
-  let inputElement3: any
+  let inputElement1: HTMLInputElement | null
+  let inputElement2: HTMLInputElement | null
+  let inputElement3: HTMLInputElement | null
 
   React.useEffect(() => {
+    // If person is not logged in, redirect to log in
     if (!props.isLoggedIn) props.history.push('/login')
   }, [props.isLoggedIn, props.history])
 
@@ -232,12 +61,13 @@ const Settings: React.FC<Props> = (props: Props) => {
     <>
       <Header title="Instellingen" />
       <BackgroundLogo />
+
       <StyledWrapper>
         <StyledTitle>Account</StyledTitle>
         <StyledContainer focus={focused1}>
           <Icon
             icon={NameIcon}
-            onClick={(): void => inputElement1.focus()}
+            onClick={(): void => inputElement1?.focus()}
             size={24}
             cursor="pointer"
           />
@@ -257,7 +87,7 @@ const Settings: React.FC<Props> = (props: Props) => {
         <StyledContainer focus={focused2}>
           <Icon
             icon={MailIcon}
-            onClick={(): void => inputElement2.focus()}
+            onClick={(): void => inputElement2?.focus()}
             size={24}
             cursor="pointer"
           />
@@ -277,7 +107,7 @@ const Settings: React.FC<Props> = (props: Props) => {
         <StyledContainer focus={focused3}>
           <Icon
             icon={PhoneIcon}
-            onClick={(): void => inputElement3.focus()}
+            onClick={(): void => inputElement3?.focus()}
             size={24}
             cursor="pointer"
           />
@@ -300,17 +130,17 @@ const Settings: React.FC<Props> = (props: Props) => {
         <StyledToggleContainer>
           <StyledToggleLabel>Notificaties</StyledToggleLabel>
           <StyledToggle>
-            <StyledCheckbox type="checkbox"></StyledCheckbox>
+            <StyledCheckbox type="checkbox" />
             <span className="slider"></span>
           </StyledToggle>
         </StyledToggleContainer>
         <StyledHorizontalRule />
-        <Link to="/help">
+        {/* <Link to="/help">
           <StyledBottomLabel>
             <Icon icon={HelpIcon} size={20} />
             <p>Help & ondersteuning</p>
           </StyledBottomLabel>
-        </Link>
+        </Link> */}
         <Link to="/terms-of-service">
           <StyledBottomLabel>
             <Icon icon={GlobeIcon} size={20} />
